+++ mm/vmscan.c
@@ -912,7 +912,7 @@ int kswapd(void *unused)
 	 * Kswapd main loop.
 	 */
 	for (;;) {
-		static int recalc = 0;
+		static long recalc = 0;
 
 		/* If needed, try to free some memory. */
 		if (inactive_shortage() || free_shortage())
